Updated to 0.5.4
[WIP] Sea-Conch item.
[WIP] New way of spawning entities, currently affects Crimson Newts, Echofins and Butterflies.
Added a new entity, The Crimson Newt. Spawns in Crimson Forests. You can lure / breed them with Crimson Roots.
Seashells are now placeable and generate in the world.
Cattails are now Waterloggable on the bottom half and spawns in Swamps and Rivers.
Fixed a bug that made Capybara's move while sleeping.

Other Fixes.